requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
googlemaps==4.10.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0